requirements.txt
asgiref==3.11.1
coverage==7.13.4
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==40.5.1
inflection==0.5.1
iniconfig==2.3.0
isort==8.0.1
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
ruff==0.15.4
sqlparse==0.5.5
uritemplate==4.2.0
setuptools==69.0.2